- load modal with card info
