+++ app/javascript/components/board/ListCard.jsx
@@ -1,16 +1,25 @@
 import React, { Component } from 'react';
 import { Link } from "react-router-dom";
+import { inject, observer } from 'mobx-react';
 
-class ListCard extends Component {
+@inject('store')
+@observer class ListCard extends Component {
   constructor(props) {
     super(props)
+
+    this.onClick = this.onClick.bind(this);
+  }
+
+  onClick(event) {
+    const { Board } = this.props.store;
+    Board.setModalCard(this.props.card);
   }
 
   render() {
     const { card } = this.props;
     return (
       <React.Fragment>
-        <Link to='#' className='list-card' data-toggle="modal" data-target="#listCardModal">
+        <Link to='#' className='list-card' data-toggle="modal" data-target="#listCardModal" onClick={this.onClick}>
           <div className="list-card-details">
             <span className="list-card-title">{card.name}</span>
             <div className="badges">
